+++ apps/builder/app/builder/features/settings-panel/props-section/props-section.stories.tsx
@@ -1,4 +1,8 @@
 import { useState } from "react";
+import type { Instance, Prop } from "@webstudio-is/project-build";
+import type { PropMeta } from "@webstudio-is/react-sdk";
+import { textVariants } from "@webstudio-is/design-system";
+import type { Asset } from "@webstudio-is/site-data";
 import { PropsSection } from "./props-section";
 import { usePropsLogic } from "./use-props-logic";
 import {
@@ -10,10 +14,6 @@ import {
   selectedPageIdStore,
 } from "~/shared/nano-states";
 import { setMockEnv } from "~/shared/env";
-import type { Instance, Prop } from "@webstudio-is/project-build";
-import type { PropMeta } from "@webstudio-is/react-sdk";
-import { textVariants } from "@webstudio-is/design-system";
-import type { Asset } from "@webstudio-is/asset-uploader";
 // eslint-disable-next-line import/no-internal-modules
 import catPath from "./props-panel.stories.assets/cat.jpg";
 
@@ -360,14 +360,14 @@ const startingProps: Prop[] = [
     instanceId,
     name: "addedUrlAttachment",
     type: "asset",
-    value: Array.from(assetsStore.get().keys() ?? [])[0] ?? "",
+    value: (Array.from(assetsStore.get().keys())[0] as string) ?? "",
   },
   {
     id: unique(),
     instanceId,
     name: "addedFile",
     type: "asset",
-    value: Array.from(assetsStore.get().keys() ?? [])[0] ?? "",
+    value: (Array.from(assetsStore.get().keys())[0] as string) ?? "",
   },
 ];
