+++ apps/builder/app/builder/features/settings-panel/shared.tsx
@@ -8,7 +8,7 @@ import {
 import equal from "fast-deep-equal";
 import type { PropMeta } from "@webstudio-is/react-sdk";
 import type { Prop } from "@webstudio-is/project-build";
-import type { Asset } from "@webstudio-is/asset-uploader";
+import type { Asset } from "@webstudio-is/site-data";
 import { SubtractIcon } from "@webstudio-is/icons";
 import {
   SmallIconButton,
