+++ apps/builder/app/builder/features/style-panel/controls/font-weight/is-supported-font-weight.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from "@jest/globals";
-import type { FontAsset } from "@webstudio-is/asset-uploader";
+import type { FontAsset } from "@webstudio-is/site-data";
 import { isSupportedFontWeight } from "./is-supported-font-weight";
 
 const createServerAsset = (meta: FontAsset["meta"]): FontAsset => ({
