+++ apps/builder/app/builder/shared/assets/use-assets.tsx
@@ -2,7 +2,8 @@ import { useMemo } from "react";
 import { useStore } from "@nanostores/react";
 import warnOnce from "warn-once";
 import store from "immerhin";
-import { type AssetType, type Asset } from "@webstudio-is/asset-uploader";
+import type { Asset } from "@webstudio-is/site-data";
+import type { AssetType } from "@webstudio-is/asset-uploader";
 import { toast } from "@webstudio-is/design-system";
 import { sanitizeS3Key } from "@webstudio-is/asset-uploader";
 import { restAssetsUploadPath, restAssetsPath } from "~/shared/router-utils";
