+++ packages/asset-uploader/src/db/load.ts
@@ -4,7 +4,7 @@ import {
   type AppContext,
   AuthorizationError,
 } from "@webstudio-is/trpc-interface/index.server";
-import type { Asset } from "../schema";
+import type { Asset } from "@webstudio-is/site-data";
 import { formatAsset } from "../utils/format-asset";
 
 export const loadAssetsByProject = async (
