+++ packages/react-sdk/package.json
@@ -37,11 +37,11 @@
   "dependencies": {
     "@jsep-plugin/assignment": "^1.2.1",
     "@nanostores/react": "^0.7.1",
-    "@webstudio-is/asset-uploader": "workspace:^",
     "@webstudio-is/css-data": "workspace:^",
     "@webstudio-is/css-engine": "workspace:^",
     "@webstudio-is/fonts": "workspace:^",
     "@webstudio-is/project-build": "workspace:^",
+    "@webstudio-is/site-data": "workspace:^",
     "html-tags": "^3.3.1",
     "jsep": "^1.3.8",
     "nanoevents": "^8.0.0",
