+++ packages/react-sdk/src/context.tsx
@@ -1,6 +1,6 @@
 import { type ReadableAtom, atom } from "nanostores";
 import { createContext } from "react";
-import type { Assets } from "@webstudio-is/asset-uploader";
+import type { Assets } from "@webstudio-is/site-data";
 import type { DataSource, Instance, Prop } from "@webstudio-is/project-build";
 import type { Pages, PropsByInstanceId } from "./props";
 import type { IndexesWithinAncestors } from "./instance-utils";
