+++ packages/react-sdk/src/tree/create-elements-tree.tsx
@@ -5,7 +5,7 @@ import {
   type ReactNode,
 } from "react";
 import type { ReadableAtom } from "nanostores";
-import type { Assets } from "@webstudio-is/asset-uploader";
+import type { Assets } from "@webstudio-is/site-data";
 import type { Instance, Instances } from "@webstudio-is/project-build";
 import type { Components } from "../components/components-utils";
 import {
