+++ packages/site-data/tsconfig.dts.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src"],
+  "compilerOptions": {
+    "declarationDir": "lib/types"
+  }
+}
